Move qos_event.py to event_handler.py.
And add a deprecation warning to qos_event.py
