+++ rclpy/test/test_qos_event.py
@@ -17,20 +17,20 @@ import unittest
 from unittest.mock import Mock
 
 import rclpy
+from rclpy.event_handler import PublisherEventCallbacks
+from rclpy.event_handler import QoSLivelinessChangedInfo
+from rclpy.event_handler import QoSLivelinessLostInfo
+from rclpy.event_handler import QoSOfferedDeadlineMissedInfo
+from rclpy.event_handler import QoSOfferedIncompatibleQoSInfo
+from rclpy.event_handler import QoSPublisherEventType
+from rclpy.event_handler import QoSRequestedDeadlineMissedInfo
+from rclpy.event_handler import QoSRequestedIncompatibleQoSInfo
+from rclpy.event_handler import QoSSubscriptionEventType
+from rclpy.event_handler import SubscriptionEventCallbacks
 from rclpy.impl.implementation_singleton import rclpy_implementation as _rclpy
 from rclpy.qos import QoSDurabilityPolicy
 from rclpy.qos import QoSPolicyKind
 from rclpy.qos import QoSProfile
-from rclpy.qos_event import PublisherEventCallbacks
-from rclpy.qos_event import QoSLivelinessChangedInfo
-from rclpy.qos_event import QoSLivelinessLostInfo
-from rclpy.qos_event import QoSOfferedDeadlineMissedInfo
-from rclpy.qos_event import QoSOfferedIncompatibleQoSInfo
-from rclpy.qos_event import QoSPublisherEventType
-from rclpy.qos_event import QoSRequestedDeadlineMissedInfo
-from rclpy.qos_event import QoSRequestedIncompatibleQoSInfo
-from rclpy.qos_event import QoSSubscriptionEventType
-from rclpy.qos_event import SubscriptionEventCallbacks
 from rclpy.task import Future
 
 from test_msgs.msg import Empty as EmptyMsg
